refactor(content): restructure into shared/cli/web layout
