+++ packages/content/src/cli/commands.ts
